fix:component re-rendering
