+++ src/components/app/Root.js
@@ -10,7 +10,7 @@ import { FooterLinks } from './FooterLinks';
 import { ErrorBoundary } from './ErrorBoundary';
 import { AppViewRouter } from './AppViewRouter';
 import { ProfileLoader } from './ProfileLoader';
-import ServiceWorkerManager from './ServiceWorkerManager';
+import { ServiceWorkerManager } from './ServiceWorkerManager';
 
 import type { Store } from 'firefox-profiler/types';
